requirements.txt
openai>=1.0.0
requests>=2.25.0
lemonade-sdk